requirements.txt
moviepy==1.0.3
pydub==0.25.1
Pillow==10.2.0
openai-whisper==20231117
numpy>=1.23
torch>=1.10.0
ffmpeg-python==0.2.0
tk